requirements.txt
streamlit>=1.23.0
yfinance
pandas
scikit-learn
numpy
python-dotenv
openai>=0.27.0

